[FIX] RESTful API로 수정

+++ src/main/java/com/quailss/demo/service/ReviewService.java
@@ -4,8 +4,10 @@ import com.quailss.demo.domain.Member;
 import com.quailss.demo.domain.Recipe;
 import com.quailss.demo.domain.Review;
 import com.quailss.demo.domain.dto.ReviewCommand;
+import com.quailss.demo.exception.EntityNotFoundException;
 import com.quailss.demo.repository.ReviewRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
@@ -16,6 +18,7 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class ReviewService {
     private final ReviewRepository reviewRepository;
+    private final RecipeService recipeService;
 
     public Optional<Review> findById(Long reviewId) {
         return reviewRepository.findById(reviewId);
@@ -29,7 +32,10 @@ public class ReviewService {
         return reviewRepository.findByMemberId(memberId);
     }
 
-    public void insertReview(Recipe recipe, Member member, BigDecimal score, String content) {
+    public void insertReview(Long recipeId, Member member, BigDecimal score, String content) {
+        Recipe recipe = recipeService.getRecipe(recipeId)
+                .orElseThrow(() -> new EntityNotFoundException.RecipeNotFoundException("레시피를 찾을 수 없습니다."));
+
         Review newReview = Review.builder()
                 .member(member)
                 .recipe(recipe)
@@ -39,11 +45,24 @@ public class ReviewService {
         reviewRepository.save(newReview);
     }
 
-    public Review updateReview(Long reviewId, ReviewCommand reviewCommand) {
+    public Review updateReview(Long reviewId, Member loggedInMember, ReviewCommand reviewCommand) {
+        Review existingReview = findById(reviewId)
+                .orElseThrow(() -> new EntityNotFoundException.ReviewNotFoundException("존재하지 않는 리뷰입니다."));
+
+        if (!existingReview.getMember().getId().equals(loggedInMember.getId())) {
+            throw new AccessDeniedException("리뷰 작성자가 아닙니다.");
+        }
+
         return reviewRepository.updateReview(reviewId, reviewCommand.getScore(), reviewCommand.getContent());
     }
 
-    public void deleteReview(Review review) {
+    public void deleteReview(Long reviewId, Member loggedInMember) {
+        Review review = findById(reviewId)
+                .orElseThrow(() -> new EntityNotFoundException.ReviewNotFoundException("존재하지 않는 리뷰입니다."));
+
+        if(!review.getMember().getEmail().equals(loggedInMember.getEmail()))
+            throw new AccessDeniedException("리뷰 작성자가 아닙니다.");
+
         reviewRepository.delete(review);
     }
 }
